syncbase.js: Inital repo and test setup.
The Makefile has tasks for running tests in both node and chrome.

The test harness starts a syncbased instance and mounts it in the test
namespace.

Currently, the tests just make RPCs directly to syncbased, mostly to
make sure the setup is working.

There is also a make task for generating syncbase vdl.js files, and
those files have been checked in to the repo in src/gen-vdl.

This CL also adds the necessary copyright headers and files.

Change-Id: I8a851539133be69660efd7df29c454af1c3ffafb