syncbase.js: Stubs for the Database and batch part of NoSqlDatabase.
Just empty stubs, nothing is currently implemented.
Note that I did not do inheritance of a common DatabaseHandler
like the Go API. Since DatabaseHandler is not used as an independent
type in the API and is pretty much internal, using inheritance in
JS would just make the documentation more confusing without giving
much benefit. If we end up exposing DatabaseHandler as a type for
some arguments or if we end up with a lot more methods on it, then
we can reconsider using inheritance.

Change-Id: I61bd3863f8dac524de54c140bfae12e9046b805f