javascript/syncbase: change js test to use "Type(v)" like idiom.
Change test to use the:
Type(v) like "%.<short-type-name>"

Support for  Type(v) = "<short-type_name>"
is being removed.  As such, the string
comparison will work like all other string
comparisons in query language expressions.

Changing this to have a different branch name than "type_idiom" as
a multi-part submission is not needed and; furthermore, seems to
be in a bad state.

Change-Id: Ia85c8335523af21799c3ce0fe3b57f435a91a078