javascript/syncbase: Allowing methods that take RowRange/PrefixRange to also accept strings. When a string is used, it is interpreted as 'prefix'.
scan and deleteRange accept string in addition
to RowRange. When a string is used, it is assumed
to be a prefix rather than a single row key.
This is a reasonable assumption since deleteRange
has the sister method delete(string key) that can
be used for deleting single rows and scanning for
a single row is equivalent to doing a get.

Now we are also removing PrefixRange completely
since prefix is now just a string for scan,
deleteRange and setPermission.

PrefixPermission also now has the prefix property
as string instead of PrefixRange.

Closes https://github.com/vanadium/issues/issues/677

Change-Id: If1c611ef6f5d0bc88ad8c2b45df2784d32ecf276